// packages/contract/src/generated/PublicReferencedSessionHead.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { SessionReference } from "./SessionReference";
import type { WireSessionHead } from "./WireSessionHead";

export type PublicReferencedSessionHead = { reference: SessionReference; session: WireSessionHead };